refactor: move LLMMessageHistoryItem to DataDefinitions and GroqApiResponse to API schemas folder
