+++ Content.Server/Imperial/XxRaay/Systems/LLMNPCSystem/GroqApiSchemas.cs
@@ -1,0 +1,21 @@
+using System.Text.Json.Serialization;
+
+namespace Content.Server.Imperial.XxRaay.Systems.LLMNPCSystem;
+
+internal sealed class GroqApiResponse
+{
+    [JsonPropertyName("choices")]
+    public GroqApiChoice[]? Choices { get; set; }
+}
+
+internal sealed class GroqApiChoice
+{
+    [JsonPropertyName("message")]
+    public GroqApiMessage? Message { get; set; }
+}
+
+internal sealed class GroqApiMessage
+{
+    [JsonPropertyName("content")]
+    public string? Content { get; set; }
+}
